Move most of SST engine tests to staging-common.  Rewrite tests to take engine and filename parameters so that they can be used by other engines.  Filename parameter helps avoid conflicts between tests.  Eliminate _nompi fortran versions in favor of conditional compilation.
