+++ testing/adios2/engine/sst/TestSstWriterFails.cpp
@@ -13,8 +13,6 @@
 
 #include <gtest/gtest.h>
 
-#include "TestData.h"
-
 class SstWriteFails : public ::testing::Test
 {
 public:
@@ -83,6 +81,8 @@ TEST_F(SstWriteFails, InvalidPut)
     io.SetParameters(NoReaders);
 
     adios2::Engine engine = io.Open(fname, adios2::Mode::Write);
+
+    double data_scalar_R64 = 0.0;
 
     EXPECT_THROW(engine.Put(scalar_r64, data_scalar_R64), std::logic_error);
