+++ testing/adios2/engine/staging-common/TestCommonWrite.cpp
@@ -15,14 +15,15 @@
 
 #include "TestData.h"
 
-class SstWriteTest : public ::testing::Test
+class CommonWriteTest : public ::testing::Test
 {
 public:
-    SstWriteTest() = default;
+    CommonWriteTest() = default;
 };
 
 adios2::Params engineParams = {}; // parsed from command line
-std::string fname = "ADIOS2Sst";
+std::string fname = "ADIOS2Common";
+std::string engine = "SST";
 
 int CompressSz = 0;
 int CompressZfp = 0;
@@ -61,8 +62,8 @@ static adios2::Params ParseEngineParams(std::string Input)
     return Ret;
 }
 
-// ADIOS2 SST write
-TEST_F(SstWriteTest, ADIOS2SstWrite)
+// ADIOS2 COMMON write
+TEST_F(CommonWriteTest, ADIOS2CommonWrite)
 {
     // form a mpiSize * Nx 1D array
     int mpiRank = 0, mpiSize = 1;
@@ -135,40 +136,8 @@ TEST_F(SstWriteTest, ADIOS2SstWrite)
         }
     }
 
-    const std::string zero = std::to_string(0);
-    const std::string s1_Single = std::string("s1_Single_") + zero;
-    const std::string s1_Array = std::string("s1_Array_") + zero;
-    const std::string i8_Single = std::string("i8_Single_") + zero;
-    const std::string i16_Single = std::string("i16_Single_") + zero;
-    const std::string i32_Single = std::string("i32_Single_") + zero;
-    const std::string i64_Single = std::string("i64_Single_") + zero;
-    const std::string u8_Single = std::string("u8_Single_") + zero;
-    const std::string u16_Single = std::string("u16_Single_") + zero;
-    const std::string u32_Single = std::string("u32_Single_") + zero;
-    const std::string u64_Single = std::string("u64_Single_") + zero;
-    const std::string r32_Single = std::string("r32_Single_") + zero;
-    const std::string r64_Single = std::string("r64_Single_") + zero;
-
-    io.DefineAttribute<std::string>(s1_Single, data_S1);
-    //        io.DefineAttribute<std::string>(s1_Array,
-    //                                        data_S1array.data(),
-    //                                        data_S1array.size());
-
-    io.DefineAttribute<int8_t>(i8_Single, data_I8.front());
-    io.DefineAttribute<int16_t>(i16_Single, data_I16.front());
-    io.DefineAttribute<int32_t>(i32_Single, data_I32.front());
-    io.DefineAttribute<int64_t>(i64_Single, data_I64.front());
-
-    io.DefineAttribute<uint8_t>(u8_Single, data_U8.front());
-    io.DefineAttribute<uint16_t>(u16_Single, data_U16.front());
-    io.DefineAttribute<uint32_t>(u32_Single, data_U32.front());
-    io.DefineAttribute<uint64_t>(u64_Single, data_U64.front());
-
-    io.DefineAttribute<float>(r32_Single, data_R32.front());
-    io.DefineAttribute<double>(r64_Single, data_R64.front());
-
     // Create the Engine
-    io.SetEngine("Sst");
+    io.SetEngine(engine);
     io.SetParameters(engineParams);
 
     adios2::Engine engine = io.Open(fname, adios2::Mode::Write);
@@ -176,7 +145,7 @@ TEST_F(SstWriteTest, ADIOS2SstWrite)
     for (size_t step = 0; step < NSteps; ++step)
     {
         // Generate test data for each process uniquely
-        generateSstTestData(step, mpiRank, mpiSize);
+        generateCommonTestData(step, mpiRank, mpiSize);
 
         engine.BeginStep();
         // Retrieve the variables that previously went out of scope
@@ -269,11 +238,31 @@ int main(int argc, char **argv)
             argv++;
             argc--;
         }
+        else if (std::string(argv[1]) == "--engine")
+        {
+            engine = std::string(argv[2]);
+            argv++;
+            argc--;
+        }
         else
         {
             throw std::invalid_argument("Unknown argument \"" +
                                         std::string(argv[1]) + "\"");
         }
+        argv++;
+        argc--;
+    }
+    if (argc > 1)
+    {
+        /* first arg without -- is engine */
+        engine = std::string(argv[1]);
+        argv++;
+        argc--;
+    }
+    if (argc > 1)
+    {
+        /* second arg without -- is filename */
+        fname = std::string(argv[1]);
         argv++;
         argc--;
     }
